+++ spec/integration/repository_spec.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
 require 'elasticsearch'
-require 'register_sources_sk/repositories/record_repository'
+require 'register_sources_sk/repository'
 require 'register_sources_sk/services/es_index_creator'
 require 'register_sources_sk/structs/record'
 
-RSpec.describe RegisterSourcesSk::Repositories::RecordRepository do
+RSpec.describe RegisterSourcesSk::Repository do
   subject { described_class.new(client: es_client, index:) }
 
-  let(:index) { SecureRandom.uuid }
+  let(:index) { "tmp-#{SecureRandom.uuid}" }
   let(:es_client) { Elasticsearch::Client.new }
 
   let(:record) do
@@ -18,11 +18,8 @@ RSpec.describe RegisterSourcesSk::Repositories::RecordRepository do
   end
 
   before do
-    index_creator = RegisterSourcesSk::Services::EsIndexCreator.new(
-      client: es_client,
-      index:
-    )
-    index_creator.create_index
+    index_creator = RegisterSourcesSk::Services::EsIndexCreator.new(client: es_client)
+    index_creator.create_index(index)
   end
 
   describe '#store' do
